% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyprtan.R
\name{hypertan}
\alias{hypertan}
\title{Hyperbolic tangent function
Hypertan function for normalization.}
\usage{
hypertan(x)
}
\arguments{
\item{x}{Input value}
}
\description{
Hyperbolic tangent function
Hypertan function for normalization.
}
\examples{
hypertan(1.5)

}
