% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{logistic_sigmoid}
\alias{logistic_sigmoid}
\title{Logistic sigmoid function
Logistic sigmoid function for normalization.}
\usage{
logistic_sigmoid(x)
}
\arguments{
\item{x}{Input value}
}
\description{
Logistic sigmoid function
Logistic sigmoid function for normalization.
}
\examples{
logistic_sigmoid(1.5)
}
